julius-api-server/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiostream==0.5.2
altair==5.3.0
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.22.0
blinker==1.8.2
boto3==1.34.162
botocore==1.34.162
browserbase==0.3.0
bs4==0.0.2
cachelib==0.9.0
cachetools==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cs50==9.0.0
cycler==0.12.1
daily-python==0.10.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
deepgram-sdk==3.2.7
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
elevenlabs==1.6.0
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
firecrawl==0.0.20
firecrawl-py==0.0.20
fireworks-ai==0.15.0
Flask==2.3.2
Flask-Caching==2.3.0
Flask-Session==0.4.0
Flask-SocketIO==5.3.6
fonttools==4.51.0
free-proxy==1.1.1
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.62.3
grpclib==0.4.7
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.14
langchain-anthropic==0.1.23
langchain-aws==0.1.17
langchain-community==0.2.12
langchain-core==0.2.34
langchain-fireworks==0.1.7
langchain-google-genai==1.0.10
langchain-google-vertexai==1.0.10
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-mistralai==0.1.12
langchain-nvidia-ai-endpoints==0.2.1
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.86
leap-python-sdk==1.1.4
loguru==0.7.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
minify_html==0.15.0
modal==0.63.41
mpire==2.10.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.42.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pipecat-ai==0.0.35
platformdirs==4.2.2
playwright==1.46.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pyloudnorm==0.1.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
scrapegraphai==1.14.0
seaborn==0.13.2
semchunk==2.2.0
sentence-transformers==3.0.1
shapely==2.0.6
shellingham==1.5.4
sigtools==4.0.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==2.1.3
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
sympy==1.12.1
synchronicity==0.6.7
tenacity==8.5.0